third_party/blink/web_tests/external/wpt/css/css-shapes/shape-outside/values/shape-outside-circle-000-expected.txt

This is a testharness.js-based test.
FAIL No arguments - inline assert_equals: expected "circle(at 50% 50%)" but got "circle()"
FAIL Radial argument only - inline assert_equals: expected "circle(50px at 50% 50%)" but got "circle(50px)"
PASS Position argument only - inline
PASS Radial and position argument - inline
FAIL No arguments - computed assert_equals: expected "circle(at 50% 50%)" but got "circle()"
FAIL Radial argument only - computed assert_equals: expected "circle(50px at 50% 50%)" but got "circle(50px)"
PASS Position argument only - computed
PASS Radial and position argument - computed
Harness: the test ran to completion.
